requirements.txt
langchain>=0.3.0
langchain-ollama>=0.2.0
langsmith>=0.1.0
pytest>=8.0.0
python-dotenv>=1.0.0
ragas
datasets